Move Node class into implementation file
