+++ inputs.cc
@@ -1,4 +1,5 @@
 
+#include <algorithm>
 #include <iostream>
 #include <map>
 #include <vector>
