+++ reader.cc
@@ -4,7 +4,30 @@
 
 #include "utils.hh"
 
+// Suppress spurious warning triggered by toml11 for gcc 4.8.5
+#if defined(__GNUC__) && __GNUC__ < 5
+#pragma GCC ignored "-Wno-unused-parameter"
+#endif
+
+#include "toml11/toml.hpp"
+
+// Re-enable warnings
+#if defined(__GNUC__) && __GNUC__ < 5
+#pragma GCC warning "-Wno-unused-parameter"
+#endif
+
+class Node {
+public:
+	typedef toml::basic_value<toml::discard_comments,
+														std::unordered_map, std::vector> value_type;
+	Node(const value_type& v) : v(v) {}
+	value_type v;
+};
+
 using namespace std;
+
+InputNode::InputNode(const Node n, const std::string& label) :
+		n_(std::shared_ptr<Node>(new Node{n})), label_(label) {}
 
 size_t InputNode::size() const
 {
